Add sha'd anon user password to secrets.
It would appear that it's going to take a large amount of juggling to
convert the anon user password from plaintext into ldap format via
puppet. Instead, we can just expect the user to produce it for us and
encode it here. It's not likely to be changed often, and we can improve
it in the future