Actually, split www hostname config a little:
After some consideration we actually have two hostnames in play: one is
the canonical hostname for the website. For example,
www.studentrobotics.org is the canonical name of our website, rather
than srobo.org or studentrobotics.org

On the other hand, we have another name in play, studentrobotics.org, on
top of which things like 'trac.' and 'git.' are built. So, allow for
configuring that.