Switch the admin's username to be anon in phpbb.
The moment that you try to switch to ldap auth, it refuses because
there's no 'admin' user to log in using ldap in the future. This is
increadibly sane and useful!

Unfortunately it's not so great to configure, because we now have to use
a user who is always going to be in ldap. Pick on anon.