Update forums dump
This is a recent dump from the live forums after they'd been cleaned
out for the SR2017 season. Admin login is done with the 'anon'
account (with password 123456).

Notes for future updates: as the database data also includes all the
configuration for the forums it includes things which need massaging
after dumping from the live site. In particular the config table
contains the following items:
- the password for the anonymous ldap user (default details above),
  stored against name 'ldap_password'
- the name of the server, stored with name 'server_name'
- the domain used for cookies, 'cookie_domain'
- various 'random' and 'salt' values which are probably not a good
  idea to make public

Change-Id: I6266b50f2033ea3265a9af854fd3777aeb40b6f7